separate login and register func in routes
